chore: Renamed definition/type/get-constructor-of.type.mts to definition/type/get-constructor-of.mts.
