+++ packages/architectura/src/core/server/server.mts
@@ -6,7 +6,7 @@ import { HTTPStatusCodeEnum } from "../../definition/enum/http-status-code.enum.
 
 import { PortsEnum } from "../../definition/enum/ports.enum.mjs";
 
-import { getConstructorOf } from "../../definition/type/get-constructor-of.type.mjs";
+import { getConstructorOf } from "../../definition/type/get-constructor-of.mjs";
 
 
 import { FileSystemService } from "../../service/file-system/file-system.service.mjs";
