+++ packages/architectura/src/definition/type/get-constructor-of.mts
